relations , 모델 배열 삭제 , backend fin

+++ src/controllers/videoController.js
@@ -1,5 +1,6 @@
 //video model 사용
 import Video from "../models/Video"
+import User from "../models/User"
 export const home = async(req, res) =>{
   const videos = await Video.find({})
   return res.render("home", { pageTitle: "Home" , videos });
@@ -8,7 +9,7 @@ export const home = async(req, res) =>{
 export const watch = async(req, res) => {
   const {id}=req.params
   //데이터베이스에서 id로 id에 해당하는 것 찾기
-  const video = await Video.findById(id)
+  const video = await Video.findById(id).populate("owner") //######## video모델 스키마에서 ref설정한걸 여기서 사용 => owner에 ref설정한 user 를 넣는다??채운다?? 
   if(!video){
     return res.render('404', {pageTitle : 'Video not found'})
   }
@@ -17,17 +18,26 @@ export const watch = async(req, res) => {
 
 export const getEdit = async(req, res) => {
   const {id}=req.params
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id)
   //에러를 항상 먼저 잡아줘라!
   if(!video){
     // 여기서 return 하지 않으면 밑에꺼도 실행되기때문에 꼭 return 해준다
     return res.status(404).render("404",{pageTitle : "Video not found."})
   }
+  if (String(video.owner) !== String(_id)) { // 본인만 수정할수있게!
+    return res.status(403).redirect("/");
+  }
   return res.render("edit",{pageTitle : `Edit ${video.title}`,video});
 };
 
 export const postEdit = async(req, res) => {
   const {id} =req.params
+  const {
+    user: { _id },
+  } = req.session;
   const {title,description,hashtags} = req.body
   console.log(title,description,hashtags)
 
@@ -36,6 +46,9 @@ export const postEdit = async(req, res) => {
   const video = await Video.exists({_id : id})
   if(!video){
     return res.status(404).render("404",{pageTitle : "Video not found."})
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   //먼저 위에서 video 가 있는지 판별하고 있으면 id를 통해 찾아서 업데이트 한다
   await Video.findByIdAndUpdate(id, {
@@ -47,19 +60,26 @@ export const postEdit = async(req, res) => {
 };
 
 
-export const getUpload = (req, res) => res.render("upload" , {pageTitle : 'Upload Video'});
+export const getUpload = (req, res) => res.render("video/upload" , {pageTitle : 'Upload Video'});
 export const postUpload = async(req,res)=>{
+  const {user : {_id}}= req.session
+  const {path : fileUrl}=req.file
   const { title, description, hashtags } = req.body;
   try {
     //데이터 베이스에 생성 및 저장
-    await Video.create({
+    const newVideo = await Video.create({
+      owner : _id,
+      fileUrl,
       title,
       description,
       hashtags: Video.formatHashTag(hashtags), 
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
-    return res.status(400).render("upload", {
+    return res.status(400).render("video/upload", {
       pageTitle: "Upload Video",
       errorMessage: error._message,
     });
@@ -69,6 +89,21 @@ export const postUpload = async(req,res)=>{
 
 export const deleteVideo = async(req, res) => {
   const {id} = req.params
+  
+  const {
+    user: { _id },
+  } = req.session;
+
+  const video = await Video.findById(id);
+  const user = await User.findById(video.owner)
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  user.videos.pull(video._id)   // user에 videos에서 삭제 되지 않는 이슈 해결
+  user.save()
   await Video.findByIdAndDelete(id)
   return res.redirect("/")
 };
